+++ SecurityService.Application/IPCameraApplication.cs
@@ -1,4 +1,5 @@
-﻿using SecurityService.Application.Service.Dtos.Camera.IPCamera;
+﻿using _0_Framework.Application;
+using SecurityService.Application.Service.Camera.IPCamera;
 using SecuritySystem.Domain.Camera;
 
 namespace SecuritySystem.Application
@@ -12,26 +13,32 @@ namespace SecuritySystem.Application
             _repository = repository;
         }
 
-        public void Create(CreateIPCamera command)
+        public OperationResult Create(CreateIPCamera command)
         {
+            var result = new OperationResult(); 
             if (_repository.Exists(x => x.HostAddress == command.HostAddress))
             {
-                return;
+                return result.Failed(ApplicationMessages.RecordNotFound);
             }
             var _create = new IPCamera(command.HostAddress, command.UserName, command.Password,command.StreamAddress,command.CameraName);
             _repository.Create(_create);
             _repository.SaveChanges();
+            return result.Succedded();
         }
 
-        public void Edit(EditIPCamera editIpc)
+        public OperationResult Edit(EditIPCamera editIpc)
         {
+            var result = new OperationResult();
+
             var _c = _repository.Get(editIpc.Id);
             if (_repository.Exists(x => x.HostAddress == editIpc.HostAddress && x.Id != editIpc.Id))
             {
-                return;
+                return result.Failed(ApplicationMessages.DuplicatedRecord);
             }
             _c.Edit(editIpc.HostAddress, editIpc.UserName, editIpc.Password, editIpc.StreamAddress,editIpc.CameraName);
             _repository.SaveChanges();
+            return result.Succedded();
+
         }
 
         public EditIPCamera GetDetails(long id)
@@ -39,10 +46,18 @@ namespace SecuritySystem.Application
             return _repository.GetDetails(id);
         }
 
-        public void Remove(long id)
+        public OperationResult Remove(long id)
         {
-            _repository.Remove(id);
-            _repository.SaveChanges();
+            var result = new OperationResult();
+            var _c = _repository.Get(id);
+            if (_repository.Exists(x => _c.Id == id))
+            {
+                _repository.Remove(id);
+                _repository.SaveChanges();
+                return result.Succedded();
+            }
+            return result.Failed(ApplicationMessages.RecordNotFound);
+
         }
 
         public List<IPCameraViewModel> Search(IPCameraSearchModel command)
